Replace uint64_t(.) with static_cast<uint64_t>(.)

+++ Common/src/Bytestream.cpp
@@ -68,7 +68,7 @@ auto readBytes(istream &stream, size_t bytes) -> uint64_t {
 auto getUint8(istream &stream) -> uint8_t { return static_cast<uint8_t>(readBytes(stream, 1)); }
 auto getUint16(istream &stream) -> uint16_t { return static_cast<uint16_t>(readBytes(stream, 2)); }
 auto getUint32(istream &stream) -> uint32_t { return static_cast<uint32_t>(readBytes(stream, 4)); }
-auto getUint64(istream &stream) -> uint64_t { return uint64_t(readBytes(stream, 8)); }
+auto getUint64(istream &stream) -> uint64_t { return readBytes(stream, 8); }
 
 auto readString(istream &stream, size_t bytes) -> string {
   auto result = string(bytes, '\0');
